refactor some store encryption stuff (#2434)
